+++ src/ElementInterface.php
@@ -77,11 +77,4 @@ interface ElementInterface
      * @return string
      */
     public function getAttribute($name);
-
-    /**
-     * @param ElementInterface $element
-     *
-     * @return bool
-     */
-    public function equals(ElementInterface $element);
 }